Avoid favicon.ico requests
When chrome fires up it requests for /favicon.ico automatically. As
chrome can start up in a tab, it won't be killed and the request can
occur during the start up of the next browser. This makes sure to only
test the user agent when the request is for the root page.